refactor(mouse): move topmost_surface_at to surface::geometry
Relocates the surface-arbitration helper next to the other surface
geometry primitives (cell_at_pane / phys_to_pane_local) it belongs
with, ahead of the broader mouse-input feature-first reorg. Its two
unit tests move with it from input::mouse to surface::geometry.
