+++ src/input/default_mode.rs
@@ -17,8 +17,8 @@ use crate::input::focus::{KeyboardDisabled, MouseDisabled};
 use crate::input::ime::{ImeCommit, ImeState};
 use crate::surface::OzmaTerminal;
 use crate::surface::geometry::phys_to_pane_local;
+use crate::surface::geometry::topmost_surface_at;
 use crate::ui::copy_mode::CopyModeState;
-use crate::webview_pointer::topmost_surface_at;
 use bevy::ecs::system::SystemParam;
 use bevy::prelude::*;
 use bevy::ui::{ComputedNode, UiGlobalTransform};
