+++ src/input/hyperlink.rs
@@ -15,8 +15,8 @@
 use crate::input::focus::MouseDisabled;
 use crate::input::{InputPhase, current_modifiers};
 use crate::surface::OzmaTerminal;
+use crate::surface::geometry::topmost_surface_at;
 use crate::surface::geometry::{cell_at_local, phys_to_pane_local};
-use crate::webview_pointer::topmost_surface_at;
 use bevy::ecs::entity::Entity;
 use bevy::input::ButtonInput;
 use bevy::input::keyboard::{KeyCode, KeyboardInput};
